Reorganize system package: move `api/system` to `internal/system` and update imports
